+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Yii Framework 2 HTTP client extension Change Log
 2.0.9 under development
 -----------------------
 
-- no changes in this release.
+- Bug #149: Fix type error in `StreamTransport` when `$http_response_header = null` (alexkart)
 
 
 2.0.8 April 16, 2019