+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Yii Framework 2 HTTP client extension Change Log
 - Bug #171: Added "Content-Length: 0" header when sending request with empty body (alexkart)
 - Bug #149: Fixed type error in `StreamTransport` when `$http_response_header = null` (alexkart)
 - Enh #66: Added `CURLOPT_FILE` option support to `CurlTransport` (alexkart)
+- Enh #85: Added `CurlFormatter` in order to support `CURLFile` for uploading files (alexkart)
 
 
 2.0.8 April 16, 2019