+++ CHANGELOG.md
@@ -1,11 +1,11 @@
 Yii Framework 2 HTTP client extension Change Log
 ================================================
 
-2.0.9 under development
------------------------
+2.0.9 April 23, 2019
+--------------------
 
-- Bug #171: Added "Content-Length: 0" header when sending request with empty body (alexkart)
 - Bug #149: Fixed type error in `StreamTransport` when `$http_response_header = null` (alexkart)
+- Bug #171: Added "Content-Length: 0" header when sending request with empty body (alexkart)
 - Enh #66: Added `CURLOPT_FILE` option support to `CurlTransport` (alexkart)
 - Enh #85: Added `CurlFormatter` in order to support `CURLFile` for uploading files (alexkart)